+++ src/main/java/codesquad/config/RouterConfig.java
@@ -1,15 +1,8 @@
 package codesquad.config;
 
-import codesquad.application.domain.User;
-import codesquad.application.repository.MyRepository;
-import codesquad.context.SessionContext;
-import codesquad.context.SessionContextManager;
 import codesquad.handler.*;
-import codesquad.http.*;
-import codesquad.template.HtmlElement;
+import codesquad.http.ContentType;
 import codesquad.template.HtmlManager;
-import codesquad.template.HtmlRoot;
-import codesquad.template.Model;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -22,9 +15,7 @@ import java.util.*;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 
-import static codesquad.handler.HandlerBuilder.url;
 import static codesquad.http.Method.GET;
-import static codesquad.http.Method.POST;
 
 /**
  * http 요청을 매핑 설정입니다.
@@ -37,143 +28,14 @@ public class RouterConfig {
 
     public Map<URLMatcher, HttpHandler> getHandlerMap() {
         handlerMap.putAll(staticResourceHandlerMap());
-        handlerMap.putAll(apiHandlerMap());
         handlerMap.putAll(templateResourceHandlerMap());
         return handlerMap;
     }
 
-    private Map<URLMatcher, HttpHandler> apiHandlerMap() {
-        Map<URLMatcher, HttpHandler> map = new HashMap<>();
-
-        get(
-                url("/registration").logic(request -> {
-                    HttpResponse response = HttpResponse.of(HttpStatus.MOVED_PERMANENTLY);
-                    response.addHeader("Location", "/registration/index.html");
-                    return response;
-                }),
-
-                url("/login").logic(request -> {
-                    HttpResponse response = HttpResponse.of(HttpStatus.MOVED_PERMANENTLY);
-                    response.addHeader("Location", "/login/index.html");
-                    return response;
-                }),
-
-                url("/logout").logic(request -> {
-                    SessionContext session = SessionContextManager.getSession(request);
-                    if (session != null) {
-                        SessionContextManager.clearContext(request);
-                    }
-                    HttpResponse response = HttpResponse.of(HttpStatus.SEE_OTHER);
-                    response.addHeader("Location", "/index.html");
-                    response.addHeader("Set-Cookie", "SID=; Max-Age=0");
-                    return response;
-                }),
-
-                url("/user/list").logic(request -> {
-                    try (InputStream input = this.getClass().getResourceAsStream("/templates/user/list.html")) {
-
-                        HttpResponse response;
-
-                        String html = new String(input.readAllBytes());
-                        HtmlRoot root = htmlManager.create(html);
-
-                        SessionContext session = SessionContextManager.getSession(request);
-                        if (session == null) {
-                            response = HttpResponse.of(HttpStatus.FOUND);
-                            response.addHeader("Location", "/login/index.html");
-                            return response;
-                        }
-
-                        Model model = new Model();
-                        model.setSession(session);
-
-                        List<User> users = MyRepository.source.findAllUser();
-
-                        HtmlElement userTableElement = root.findById("user-table");
-
-                        for (User user : users) {
-                            userTableElement.addChild(HtmlElement.create("<tr class=\"myclass\">")
-                                    .addChildren(
-                                            HtmlElement.create("<td>").addChildren(HtmlElement.create(user.getUserId())),
-                                            HtmlElement.create("<td>").addChildren(HtmlElement.create(user.getNickname()))
-                                    )
-                                    .setClose()
-                                    .build()
-                            );
-                        }
-
-                        root.applyModel(model);
-
-                        response = HttpResponse.of(HttpStatus.OK);
-                        response.setContentType(ContentType.TEXT_HTML);
-                        response.setBody(root.toHtml().getBytes());
-
-                        return response;
-                    } catch (IOException e) {
-                        throw new RuntimeException(e);
-                    }
-                })
-        );
-
-        post(
-                url("/user/create")
-                        .logic(request -> {
-                            HttpResponse response;
-                            if (!(
-                                    request.method.equals(POST) &&
-                                    request.getHeader("Content-Type").contains("application/x-www-form-urlencoded")
-                            )) {
-                                response = HttpResponse.of(HttpStatus.BAD_REQUEST);
-                                return response;
-                            }
-
-                            String userId = readSingleBodyParam(request, "userId");
-                            String password = readSingleBodyParam(request, "password");
-                            String nickname = readSingleBodyParam(request, "nickname");
-
-                            // save user
-                            User user = new User(userId, password, nickname);
-                            MyRepository.source.save(userId, user);
-
-                            response = HttpResponse.of(HttpStatus.SEE_OTHER);
-                            response.addHeader("Location", "/index.html");
-
-                            logger.debug("User created: {}", user);
-                            return response;
-                        }),
-                url("/signin").logic(request -> {
-                    HttpResponse response;
-                    if (!(
-                            request.method.equals(Method.POST) &&
-                            request.getHeader("Content-Type").contains("application/x-www-form-urlencoded")
-                    )) {
-                        response = HttpResponse.of(HttpStatus.BAD_REQUEST);
-                        return response;
-                    }
-
-                    String userId = readSingleBodyParam(request, "userId");
-                    String password = readSingleBodyParam(request, "password");
-
-                    User user = MyRepository.source.findUser(userId);
-                    if (user != null && user.getPassword().equals(password)) {
-
-                        String sid = SessionContextManager.createContext();
-                        SessionContext context = SessionContextManager.getContext(sid);
-                        context.setAttributes("user", user);
-
-                        response = HttpResponse.of(HttpStatus.SEE_OTHER);
-                        response.addHeader("Location", "/index.html");
-                        response.addHeader("Set-Cookie", "SID=" + sid);
-                    } else {
-                        response = HttpResponse.of(HttpStatus.SEE_OTHER);
-                        response.addHeader("Location", "/login/fail.html");
-                    }
-
-                    return response;
-                })
-        );
-
-        return map;
+    public void setRoute(List<RouteEntry> route) {
+        for (RouteEntry entry : route) {
+            handlerMap.put(entry.urlMatcher(), entry.httpHandler());
+        }
     }
 
     private Map<URLMatcher, HttpHandler> staticResourceHandlerMap() {
@@ -324,12 +186,6 @@ public class RouterConfig {
         }
     }
 
-    private String readSingleBodyParam(HttpRequest request, String attr) {
-        Object bodyParam = request.getBodyParam(attr);
-        if (bodyParam instanceof String) return (String) bodyParam;
-        return null;
-    }
-
     private ContentType getContentType(String fileName) {
         // extract file extension
         int dotIndex = fileName.lastIndexOf('.');
@@ -340,22 +196,5 @@ public class RouterConfig {
             case -1 -> ContentType.TEXT_PLAIN;
             default -> ContentType.of(extension);
         };
-    }
-
-    public RouterConfig get(HandlerBuilder... builders) {
-        return method(new Method[]{GET}, builders);
-    }
-
-    public RouterConfig post(HandlerBuilder... builders) {
-        return method(new Method[]{POST}, builders);
-    }
-
-    public RouterConfig method(Method[] methods, HandlerBuilder... builders) {
-        for (HandlerBuilder builder : builders) {
-            URLMatcher urlMatcher = URLMatcher.method(methods).urlTemplate(builder.url()).build();
-            HttpHandler httpHandler = builder.handler();
-            handlerMap.put(urlMatcher, httpHandler);
-        }
-        return this;
     }
 }
