+++ src/main/java/codesquad/context/ApplicationContext.java
@@ -2,6 +2,8 @@ package codesquad.context;
 
 import codesquad.config.FilterConfig;
 import codesquad.config.RouterConfig;
+import codesquad.handler.LoginTable;
+import codesquad.template.HtmlManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -13,10 +15,13 @@ public class ApplicationContext {
     static {
         context.getRouterConfig();
         context.getFilterConfig();
+        context.getHtmlManager();
+
+        // router init
+        context.getLoginHandler();
     }
 
     private FilterConfig filterConfig;
-    private RouterConfig routerConfig;
 
     public FilterConfig getFilterConfig() {
         if (filterConfig == null) {
@@ -26,11 +31,32 @@ public class ApplicationContext {
         return filterConfig;
     }
 
+    private RouterConfig routerConfig;
+
     public RouterConfig getRouterConfig() {
         if (routerConfig == null) {
             logger.info("init router config");
             routerConfig = new RouterConfig();
         }
         return new RouterConfig();
+    }
+
+    private HtmlManager htmlManager;
+    private LoginTable loginHandler;
+
+    public HtmlManager getHtmlManager() {
+        if (htmlManager == null) {
+            logger.info("init html manager");
+            htmlManager = new HtmlManager();
+        }
+        return htmlManager;
+    }
+
+    public LoginTable getLoginHandler() {
+        if (loginHandler == null) {
+            logger.info("init login handler");
+            loginHandler = new LoginTable(getRouterConfig(), getHtmlManager());
+        }
+        return loginHandler;
     }
 }
