+++ src/main/java/codesquad/handler/HandlerBuilder.java
@@ -1,29 +1,37 @@
 package codesquad.handler;
 
+import codesquad.http.Method;
+
 public class HandlerBuilder {
 
+    private Method[] methods;
     private String urlTemplate;
     private HttpHandler logic;
 
     private HandlerBuilder() {
     }
 
-    public static HandlerBuilder url(String urlTemplate) {
+    public static HandlerBuilder method(Method... methods) {
         HandlerBuilder handlerBuilder = new HandlerBuilder();
-        handlerBuilder.urlTemplate = urlTemplate;
+        handlerBuilder.methods = methods;
         return handlerBuilder;
     }
 
-    public String url() {
-        return urlTemplate;
+    public static HandlerBuilder get(String urlTemplate) {
+        return method(Method.GET).url(urlTemplate);
     }
 
-    public HttpHandler handler() {
-        return logic;
+    public static HandlerBuilder post(String urlTemplate) {
+        return method(Method.POST).url(urlTemplate);
     }
 
-    public HandlerBuilder logic(HttpHandler handler) {
+    public RouteEntry logic(HttpHandler handler) {
         this.logic = handler;
+        return new RouteEntry(URLMatcher.method(methods).urlTemplate(urlTemplate).build(), logic);
+    }
+
+    public HandlerBuilder url(String urlTemplate) {
+        this.urlTemplate = urlTemplate;
         return this;
     }
 }
