+++ src/main/java/codesquad/handler/RouteEntry.java
@@ -1,0 +1,7 @@
+package codesquad.handler;
+
+public record RouteEntry(
+        URLMatcher urlMatcher,
+        HttpHandler httpHandler
+) {
+}
